Move test to proper place.
